requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.5
django-cleanup==5.1.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-api-key==2.0.0
djoser==2.0.5
geographiclib==1.50
geopy==2.0.0
google==3.0.0
google-auth==1.20.1
google-oauth==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
mysqlclient==1.4.6
numpy==1.19.3
Pillow==7.1.2
proglog==0.1.9
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.0
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.55
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.48.2
urllib3==1.25.10
validate-email==1.3
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.4
